requirements.txt
torch @ https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
transformers==3.3.1
sentence-transformers==0.3.8
pandas==1.1.2
faiss-cpu==1.6.1
numpy==1.19.2
folium==0.2.1
streamlit==0.62.0
Flask==1.1.2
-e .
